requirements.txt
MouseInfo==0.1.3
numpy==2.3.1
pillow==11.3.0
PyAutoGUI==0.9.54
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
scipy==1.16.0
